Breaks TabState down in to TabCreator and TabsPagerListener to separate concerns.
