+++ landing-strip/core/src/main/java/com/novoda/landingstrip/ScrollingLandingStrip.java
@@ -13,7 +13,7 @@ public class ScrollingLandingStrip extends HorizontalScrollView implements Scrol
 
     private final Attributes attributes;
     private final Paint indicatorPaint;
-    private TabsState tabsState;
+    private TabsPagerListener tabsPagerListener;
 
     public ScrollingLandingStrip(Context context, AttributeSet attrs) {
         super(context, attrs);
@@ -23,7 +23,7 @@ public class ScrollingLandingStrip extends HorizontalScrollView implements Scrol
 
         attributes = Attributes.readAttributes(context, attrs);
         TabsContainer tabsContainer = DynamicTabsContainer.newInstance(context, attributes, attrs);
-        tabsState = TabsState.newInstance(tabsContainer, attributes, this, this);
+        tabsPagerListener = TabsPagerListener.newInstance(tabsContainer, attributes, this, this);
         this.indicatorPaint = new Paint();
         indicatorPaint.setAntiAlias(true);
         indicatorPaint.setStyle(Paint.Style.FILL);
@@ -31,32 +31,32 @@ public class ScrollingLandingStrip extends HorizontalScrollView implements Scrol
     }
 
     public void setOnPageChangeListener(ViewPager.OnPageChangeListener delegateOnPageChangeListener) {
-        tabsState.setOnPageChangeListener(delegateOnPageChangeListener);
+        tabsPagerListener.setOnPageChangeListener(delegateOnPageChangeListener);
     }
 
     @Override
     public void attach(ViewPager viewPager) {
-        tabsState.attach(viewPager);
+        tabsPagerListener.attach(viewPager);
     }
 
     @Override
     public void attach(ViewPager viewPager, PagerAdapter pagerAdapter) {
-        tabsState.attach(viewPager, pagerAdapter);
+        tabsPagerListener.attach(viewPager, pagerAdapter);
     }
 
     @Override
     public void attach(ViewPager viewPager, PagerAdapter pagerAdapter, TabSetterUpper tabSetterUpper) {
-        tabsState.attach(viewPager, pagerAdapter, tabSetterUpper);
+        tabsPagerListener.attach(viewPager, pagerAdapter, tabSetterUpper);
     }
 
     @Override
     protected void onDraw(@NonNull Canvas canvas) {
         super.onDraw(canvas);
-        if (tabsState.isEmpty()) {
+        if (tabsPagerListener.isEmpty()) {
             return;
         }
 
-        drawIndicator(canvas, tabsState.calculateIndicatorCoordinates());
+        drawIndicator(canvas, tabsPagerListener.calculateIndicatorCoordinates());
     }
 
     protected void drawIndicator(Canvas canvas, Coordinates indicatorCoordinates) {
@@ -84,12 +84,12 @@ public class ScrollingLandingStrip extends HorizontalScrollView implements Scrol
     @Override
     protected void onAttachedToWindow() {
         super.onAttachedToWindow();
-        tabsState.reregister();
+        tabsPagerListener.reregister();
     }
 
     @Override
     protected void onDetachedFromWindow() {
-        tabsState.unregister();
+        tabsPagerListener.unregister();
         super.onDetachedFromWindow();
     }
 }
